[front/lib/api/assistant/configuration] - refactor: extract listing skill versions to SkillResource
 - Centralize the logic of listing skill versions to the SkillResource class
 - Remove duplicate code from skill configuration API by leveraging new method in SkillResource
